Issue warnings for itunes tags which appear to contain HTML.
Note: this will also cause warnings to appear for titles and a
few other select tags (not descriptions!).  Previously, only
informational messages (which, by default, are not displayed)
were generated.

If this is a problem, we can change some individual tags, or
split this into two messages (one a warning, one informational).


git-svn-id: http://feedvalidator.googlecode.com/svn/trunk/feedvalidator@238 34a10a33-d82d-0410-ba40-81c901463ecc